# backend/docs/swagger.yaml
definitions:
  github_com_lshc_contract-hub_backend_pkg_response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_handler.SaveVersionRequest:
    properties:
      change_summary:
        type: string
      document_content:
        additionalProperties: true
        type: object
    type: object
  internal_handler.createShareRequest:
    properties:
      expire_hours:
        example: 24
        type: integer
      permission:
        example: 1
        type: integer
    type: object
  internal_handler.joinShareRequest:
    properties:
      name:
        example: 张三
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 心智协同合同协作系统 V1 后端接口文档
  title: 心智协同合同协作系统 API
  version: "1.0"
paths:
...
  /api/contracts/{id}/confirm:
    post:
      description: 内部用户确认当前版本
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 确认成功
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "404":
          description: 合同不存在或无权访问
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 内部用户确认合同
      tags:
      - 确认
  /api/contracts/{id}/confirmations:
    get:
      description: 查看合同的确认记录
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 确认记录
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 查看确认记录
      tags:
      - 确认
...
  /api/contracts/{id}/share:
    post:
      consumes:
      - application/json
      description: 为合同生成外部协作者分享链接
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分享参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createShareRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 分享链接
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "404":
          description: 合同不存在或无权访问
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 创建分享链接
      tags:
      - 分享协作
  /api/contracts/{id}/share/{shareId}/disable:
    post:
      description: 使分享链接失效
      parameters:
      - description: 合同ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分享记录ID
        in: path
        name: shareId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 已失效
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 失效分享链接
      tags:
      - 分享协作
...
  /api/share/{token}:
    get:
      description: 外部用户通过 token 获取合同概要
      parameters:
      - description: 分享Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分享概要
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "404":
          description: 分享链接不存在或已失效
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      summary: 获取分享概要
      tags:
      - 分享协作
  /api/share/{token}/changes:
    get:
      description: 外部协作者查看变更记录
      parameters:
      - description: 分享Token
        in: path
        name: token
        required: true
        type: string
      - description: 协作者姓名
        in: query
        name: collaborator_name
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 变更记录
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      summary: 外部变更记录
      tags:
      - 分享协作
  /api/share/{token}/confirm:
    post:
      consumes:
      - application/json
      description: 外部协作者确认当前版本
      parameters:
      - description: 分享Token
        in: path
        name: token
        required: true
        type: string
      - description: 协作者姓名
        in: query
        name: collaborator_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 确认成功
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      summary: 外部协作者确认合同
      tags:
      - 确认
  /api/share/{token}/confirmations:
    get:
      description: 外部协作者查看确认记录
      parameters:
      - description: 分享Token
        in: path
        name: token
        required: true
        type: string
      - description: 协作者姓名
        in: query
        name: collaborator_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 确认记录
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      summary: 外部查看确认记录
      tags:
      - 确认
  /api/share/{token}/contract:
    get:
      description: 外部协作者获取合同当前版本内容
      parameters:
      - description: 分享Token
        in: path
        name: token
        required: true
        type: string
      - description: 协作者姓名
        in: query
        name: collaborator_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 合同内容
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "403":
          description: 协作者不存在
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      summary: 外部获取合同内容
      tags:
      - 分享协作
...
  /api/share/{token}/join:
    post:
      consumes:
      - application/json
      description: 外部用户输入姓名加入合同协作
      parameters:
      - description: 分享Token
        in: path
        name: token
        required: true
        type: string
      - description: 姓名
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.joinShareRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 协作者信息
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      summary: 外部协作者加入
      tags:
      - 分享协作
  /api/share/{token}/versions:
    get:
      description: 外部协作者分页查看合同版本列表
      parameters:
      - description: 分享Token
        in: path
        name: token
        required: true
        type: string
      - description: 协作者姓名
        in: query
        name: collaborator_name
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 版本列表
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      summary: 外部版本列表
      tags:
      - 分享协作
    post:
      consumes:
      - application/json
      description: 外部协作者提交修改，生成新版本
      parameters:
      - description: 分享Token
        in: path
        name: token
        required: true
        type: string
      - description: 协作者姓名
        in: query
        name: collaborator_name
        required: true
        type: string
      - description: 新版本内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SaveVersionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 保存成功
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "403":
          description: 只读链接不可编辑
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      summary: 外部保存版本
      tags:
      - 分享协作
  /api/share/{token}/versions/{versionId}:
    get:
      description: 外部协作者查看指定版本详情
      parameters:
      - description: 分享Token
        in: path
        name: token
        required: true
        type: string
      - description: 版本ID
        in: path
        name: versionId
        required: true
        type: integer
      - description: 协作者姓名
        in: query
        name: collaborator_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 版本详情
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      summary: 外部版本详情
      tags:
      - 分享协作
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"